Added projects managements, updated model management

+++ api/views/LogicalModels.py
@@ -5,35 +5,45 @@ from rest_framework.exceptions import PermissionDenied
 
 from django.http import Http404
 
-from api.models import LogicalModel
+from api.models import LogicalModel, Project
 from api.serializers import LogicalModelSerializer
 
 
 class LogicalModels(APIView):
 
-	def get(self, request, format=None):
+	def get(self, request, pk):
 
 		if request.user.is_anonymous:
 			raise PermissionDenied
 
-		models = LogicalModel.objects.filter(user=request.user)
-		serializer = LogicalModelSerializer(models, many=True)
+		if Project.objects.filter(user=request.user, id=pk).exists():
 
-		return Response(serializer.data)
+			models = LogicalModel.objects.filter(
+				project=Project.objects.get(user=request.user, id=pk)
+			)
+			serializer = LogicalModelSerializer(models, many=True)
 
+			return Response(serializer.data)
+		else:
+			raise Http404
 
 	def post(self, request):
 
 		if request.user.is_anonymous:
 			raise PermissionDenied
 
-		LogicalModel(
-			user=request.user,
-			name=request.data['name'],
-			file=request.data['file']
-		).save()
+		if Project.objects.filter(user=request.user, id=int(request.data['project'])).exists():
 
-		return Response(status=status.HTTP_200_OK)
+			LogicalModel(
+				project=Project.objects.get(user=request.user, id=int(request.data['project'])),
+				name=request.data['name'],
+				file=request.data['file']
+			).save()
+
+			return Response(status=status.HTTP_200_OK)
+
+		else:
+			raise Http404
 
 
 	def delete(self, request, pk=None, format=None):
@@ -41,8 +51,15 @@ class LogicalModels(APIView):
 		if request.user.is_anonymous:
 			raise PermissionDenied
 		try:
-			model = LogicalModel.objects.get(user=request.user, id=request.data['id'])
-			model.delete()
+			if Project.objects.filter(user=request.user, id=int(request.data['project'])).exists():
+
+				model = LogicalModel.objects.get(
+					project=Project.objects.get(user=request.user, id=int(request.data['project'])),
+					id=request.data['id']
+				)
+				model.delete()
+			else:
+				raise Http404
 
 		except LogicalModel.DoesNotExist:
 			raise Http404
